requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.109.2
h11==0.14.0
idna==3.6
iso8601==1.1.0
pydantic==2.6.1
pydantic_core==2.16.2
pypika-tortoise==0.1.6
pytz==2024.1
sniffio==1.3.0
starlette==0.36.3
tortoise-orm==0.20.0
typing_extensions==4.9.0
uvicorn==0.27.1
